Add files via upload
Credit Card Fraud detction using Autoencoders in H2O.
Machine Learning PLatform used in here is H2O, which is a Fast, Scalable, Open source application for machine/deep learning.The speciality of H2O is that it is using in-memory compression to handles billions of data rows in memory, even in a small cluster. It is easy to use APIs with R, Python, Scala, Java, JSON as well as a built in web interface.